src/requirements.txt
google-generativeai>=0.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
